Error handling & Original Templates
Added error handling in many places to help the user understand what went wrong with their XML. Also added more error logging and an XML specific error log to help errors be relevant to the user.

The intention of the standard error log is that they can share it with me in the event hey belive there is a bug in the program, not their data.